+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.4.0] - 2025-04-05
+### Added / Changed / Fixed
+- Changed: function drag_mouse to drag_mouse_to_pos. Reason is the new drag_mouse function that moves mouse in relation to current position
+- made MouseClick enum public
+- Added functions: 
+-    - move_mouse() - Moves mouse relative to its current position. -x left, +x right, -y up, +y down
+-    - move_mouse_to() -- Accepts Option<x>, Option<y>. None value keeps the same x or y. Usefull for horizontal and vertical movement
+-    - drag_mouse() - performs mouse drag action relative to current position. Same rules as in move_mouse function
+-    - drag_mouse_to() - same as in mouse_move_to, accepts Options. 
+-    - drag_mouse_to_pos() - same as in move_mouse_to_pos(). This is the old drag_mouse() func 
+-    - get_mouse_position() - returns Result<(i32, i32)> current mouse position
+-    - click() - Mouse click - choose button with MouseClick enum
+-    - click_down() - accepts MouseClick enum (does not work on macOS)
+-    - click_up() - accepts MouseClick enum (does not work on macOS)
+-    - key_down() - executes key press down only 
+-    - key_up() - executes key press up only
+- move_mouse_to_pos() remains the same, while drag_mouse_to_pos() is new name for the old version of drag_mouse() function
+
 ## [2.3.0] - 2025-03-30
 ### Fixed / Removed
 - Rework of Segmented NCC template match. Completely removed argument of Max segments and made it always work robustly, never sacrificing precision for  speed. Additionally fixed part of formulas which will additionally reduce false positives, regardless of max segments. The fix also improves algorithms speed when compared to previous version, if max_segments is not taken into consideration. The speed gain is due to much less checks and verifications in the algorithm.